+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### [2.3.0] - 2022-07-
 #### Added
 - add support for deletions from `Editops`/`Opcodes`
+- add `as_matching_blocks` to `Editops`/`Opcodes`
 
 #### Changed
 - merge adjacent similar blocks in `Opcodes`